Add: Damage Calculation
#124 #110

+ 관통, 충격에 따른 데미지 계산 및 적용 로직 구현

+++ TotheNorth/Assets/Scripts/Battles/Objects/AttackInfo.cs
@@ -8,8 +8,8 @@ namespace Assets.Scripts.Items
     /// 
     public class AttackInfo
     {
-        public float powerPenetration;
-        public float powerImpact;
+        public int powerPenetration;
+        public int powerImpact;
         public float powerKnockback;
 
         public int damagePenetration, damageImpact;
